requirements.txt
streamlit
pandas
numpy
requests
beautifulsoup4
plotly
PyGithub